consolidated tests for `generate_rct_data`, `generate_obs_data`, and `CausalData.get_df` into a unified `test_data.py` for improved maintainability; renamed `causaldata` class to `CausalData` with stricter validations; updated imports and docs accordingly
