+++ docs/user-guide.rst
@@ -1,9 +1,10 @@
 User Guide
 ==========
 
-This guide will help you get started with CausalKit and understand its core concepts.
+This guide will help you get started with Causal Inference and CausalKit and understand its core concepts.
 
 .. toctree::
    :maxdepth: 2
 
+   user-guide/causal_inference_fundamentals
    user-guide/causaldata
