+++ docs/user-guide/causal_inference_fundamentals.md
@@ -1,0 +1,8 @@
+# Causal Inference Intuition
+example with pills
+# ITE
+# ATE and ATT
+What is treatment
+* What is target
+* What is cofounders
+* RCT vs Observational data. Write about ladder
